Move demo apps into a separate module
Also:
* remove actor system dependency from the demo apps - it wasn't really needed
