--- akka/tests/src/test/resources/AppendReplicate.conf
@@ -1,24 +1,0 @@
-include "application"
-
-akka {
-  loglevel = "DEBUG"
-}
-
-evolutiongaming.kafka-journal {
-
-  persistence.journal {
-    kafka {
-      bootstrap-servers = [
-        "localhost:9092",
-        "localhost:9093",
-        "localhost:9094"]
-    }
-
-    cassandra = ${evolutiongaming.kafka-journal.cassandra}
-  }
-
-  replicator {
-    kafka = ${evolutiongaming.kafka-journal.persistence.journal.kafka}
-    cassandra = ${evolutiongaming.kafka-journal.persistence.journal.cassandra}
-  }
-}
