+++ demo/src/main/resources/AppendReplicateApp.conf
@@ -1,0 +1,38 @@
+# AppendReplicateApp settings
+
+# msokolov: I have no idea why these specific overrides are needed, could be removed?
+
+evolutiongaming.kafka-journal {
+
+  persistence.journal {
+    kafka {
+      bootstrap-servers = [
+        "localhost:9092",
+        "localhost:9093",
+        "localhost:9094",
+      ]
+
+      producer {
+        max-request-size = 8388608
+      }
+    }
+
+    cassandra {
+      contact-points = ["0.0.0.0"]
+      client {
+        reconnection {
+          min-delay = 300ms
+          max-delay = 1s
+        }
+        query.consistency = "ONE"
+      }
+    }
+  }
+
+  replicator {
+    kafka = ${evolutiongaming.kafka-journal.persistence.journal.kafka}
+    cassandra = ${evolutiongaming.kafka-journal.persistence.journal.cassandra}
+
+    topic-discovery-interval = 1s
+  }
+}
